updated react and react-dom to latest version, cannot read props of undefined error
