+++ package.json
@@ -20,9 +20,10 @@
     "getusermedia": "^2.0.0",
     "isomorphic-fetch": "^2.2.1",
     "material-ui": "^0.16.6",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-redux": "^4.4.5",
+    "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
     "rtc-media": "^2.0.1",
