+++ src/index.js
@@ -5,6 +5,7 @@ import Root from './components/Root'
 
 import './index.css'
 
+
 ReactDOM.render(
   <Root />,
   document.getElementById('root')
